+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Archiving of routines
 - Button for saving changes during guided training session
 - Settings for hiding UI elements related to RPE and TUT
+- Option to prefer exercise in training session
 
 ### Changed